Split test_categorical into subpackage (#18497)
