+++ setup.py
@@ -747,6 +747,7 @@ setup(name=DISTNAME,
                 'pandas.tests',
                 'pandas.tests.api',
                 'pandas.tests.dtypes',
+                'pandas.tests.categorical',
                 'pandas.tests.computation',
                 'pandas.tests.sparse',
                 'pandas.tests.frame',
